+++ insitu/views/base.py
@@ -83,7 +83,10 @@ class ESDatatableView(BaseDatatableView, ProtectedView):
         search_text = self.request.GET.get("search[value]", "")
         if search_text:
             search = search.query(
-                "query_string", default_field="name", query='"' + search_text + '"'
+                "query_string",
+                fields=["id", "name", "acronym"],
+                query='"' + search_text + '"',
+                lenient=True,
             )
 
         if search.count() > settings.MAX_RESULT_WINDOW or not hasattr(